Valida objetos antes de salvar

+++ src/controllers/turmaController.ts
@@ -1,6 +1,12 @@
 import { Request, Response } from 'express'
 import { validationResult } from 'express-validator'
+import DisciplinaRepository from '../repositories/disciplinaRepository'
+import TurmaRepository from '../repositories/turmaRepository'
+import UsuarioRepository from '../repositories/usuarioRepository'
+import DisciplinaService from '../services/disciplinaService'
 import TurmaService from '../services/turmaService'
+import UsuarioService from '../services/usuarioService'
+import UsuarioController from './usuarioController'
 
 
 export default class TurmaController {
@@ -26,6 +32,21 @@ export default class TurmaController {
       if (!errors.isEmpty()) {
         return res.status(400).json({ errors: errors.array() })
       }
+      const { disciplinaId, professorId } = req.body
+
+      const disciplina = await DisciplinaRepository.findOneById(disciplinaId)
+      if (!disciplina) {
+        return res.status(404).json({ error: 'Disciplina não encontrada!' })
+      }
+
+      const professor = await UsuarioRepository.findOneById(professorId)
+      if (!professor) {
+        return res.status(404).json({ error: 'Professor não encontrado!' })
+      }
+      if (professor.tipo !== 'professor') {
+        return res.status(400).json({ error: 'Usuário informado não é "professor"!' })
+      }
+
       const turma = await TurmaService.post(req.body)
       return res.status(201).json(turma)
     } catch (error) {
@@ -42,7 +63,31 @@ export default class TurmaController {
         return res.status(400).json({ errors: errors.array() })
       }
       const id = req.params.id as string
-      const turma = await TurmaService.update(id, req.body)
+      let turma = await TurmaRepository.findOneById(id)
+      if (!turma) {
+        return res.status(404).json({ error: 'Turma não encontrada!' })
+      }
+
+      const { disciplinaId, professorId } = req.body
+      if (disciplinaId) {
+        const disciplina = await DisciplinaRepository.findOneById(disciplinaId)
+        if (!disciplina) {
+          return res.status(404).json({ error: 'Disciplina não encontrada!' })
+        }
+      }
+
+      if (professorId) {
+        const professor = await UsuarioRepository.findOneById(professorId)
+        if (!professor) {
+          return res.status(404).json({ error: 'Professor não encontrado!' })
+        }
+        if (professor.tipo !== 'professor') {
+          return res.status(400).json({ error: 'Usuário informado não é "professor"!' })
+        }
+      }
+
+
+      turma = await TurmaService.update(id, req.body)
       return res.status(200).json(turma)
     } catch (error) {
       console.error(error)
@@ -58,6 +103,10 @@ export default class TurmaController {
         return res.status(400).json({ errors: errors.array() })
       }
       const id = req.params.id as string
+      let turma = await TurmaRepository.findOneById(id)
+      if (!turma) {
+        return res.status(404).json({ error: 'Turma não encontrada!' })
+      }
       await TurmaService.delete(id)
       return res.sendStatus(200)
     } catch (error) {
@@ -74,7 +123,25 @@ export default class TurmaController {
         return res.status(400).json({ errors: errors.array() })
       }
       const id = req.params.id as string
+      let turma = await TurmaRepository.findOneById(id)
+      if (!turma) {
+        return res.status(404).json({ error: 'Turma não encontrada!' })
+      }
+
       const alunosId: string[] = req.body.alunosId
+      let alunoErrors = false
+      const promises = alunosId.map((alunoId) => {
+        return UsuarioRepository.findOneById(alunoId)
+      });
+
+      Promise.all(promises).then((values) => {
+        alunoErrors = values.some((aluno) => { (!aluno || aluno.tipo !== 'aluno') })
+      })
+
+      if (alunoErrors) {
+        return res.status(400).json({ error: 'Um ou mais aluno inválido!' })
+      }
+
       const alunos = await TurmaService.postAlunos(id, alunosId)
       return res.status(201).json(alunos)
     } catch (error) {
@@ -91,6 +158,11 @@ export default class TurmaController {
         return res.status(400).json({ errors: errors.array() })
       }
       const id = req.params.id as string
+      let turma = await TurmaRepository.findOneById(id)
+      if (!turma) {
+        return res.status(404).json({ error: 'Turma não encontrada!' })
+      }
+
       const result = await TurmaService.getAlunos(id)
       return res.status(200).json(result)
     } catch (error) {
